Change Int to Float and Int to Date necessary types

+++ src/modules/offer-request/dto/creation/offer-request-creation.args.ts
@@ -1,4 +1,4 @@
-import { ArgsType, Field, Int } from '@nestjs/graphql';
+import { ArgsType, Field, Float } from '@nestjs/graphql';
 import { OfferRequestDataInput } from './offer-request-data.input';
 import { OfferRequestSubscriptionInput } from './offer-request-subscription.input';
 
@@ -10,12 +10,12 @@ export class OfferRequestCreationArgs {
   @Field(() => OfferRequestSubscriptionInput, { nullable: true })
   subscription?: OfferRequestSubscriptionInput;
 
-  @Field(() => Int, { nullable: true })
+  @Field(() => Float, { nullable: true })
   customerVat?: number;
 
-  @Field(() => Int, { nullable: true })
+  @Field(() => Float, { nullable: true })
   minimumDuration?: number;
 
-  @Field(() => Int, { nullable: true })
+  @Field(() => Float, { nullable: true })
   serviceProviderVAT?: number;
 }
